2017.6.13下班

+++ 绿茵荟/MyFile/Classes/Sections/Main/2.MainSection/2.Wallpaper/WallpaperShow/Controller/WallpaperShowViewController.h
@@ -12,7 +12,7 @@
 @property(nonatomic,assign)UIImage * image_show;//要显示的image
 @property(nonatomic,copy)NSString * name;
 @property(nonatomic,strong)NSArray * img_array;//上个页面传过来的数组
-@property(nonatomic,assign)int current_index;//显示的图片在数组中的序号
+@property(nonatomic,assign)NSInteger current_index;//显示的图片在数组中的序号
 
 
 @property(nonatomic,strong)NSArray * wallpaperList;//上个页面加载界面的数据
